# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "RotaTour API",
        "description": "Documentation created using Swagger.",
        "contact": {
            "name": "Saulo Gomes",
            "url": "https://github.com/saulobr88"
        },
        "version": "1.0"
    },
    "host": "rotatourapi.herokuapp.com",
    "basePath": "",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "login"
                ],
                "summary": "get a JWT token for a user.",
                "description": "get a JWT token for a user.",
                "operationId": "api.auth",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "email",
                        "in": "body",
                        "description": "Email used by user",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/NewUser"
                        },
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "body",
                        "description": "Password used by user",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/NewUser"
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success - respond with a JWT Token"
                    },
                    "401": {
                        "description": "Invalid Credentials."
                    },
                    "500": {
                        "description": "Could not create token"
                    }
                }
            }
        },
        "/api/getuser": {
            "get": {
                "tags": [
                    "login"
                ],
                "summary": "Get the user by token - For Tests only.",
                "description": "get a user by JWT Token - for tests only.",
                "operationId": "api.getuser",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Email used by user",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success - respond with a JWT Token"
                    },
                    "404": {
                        "description": "User not found."
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "login"
                ],
                "summary": "Handle a registration request for the application.",
                "description": "Handle a registration request for the application",
                "operationId": "api.register",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "name",
                        "in": "body",
                        "description": "Fullname used by user",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/NewUser"
                        },
                        "type": "string"
                    },
                    {
                        "name": "email",
                        "in": "body",
                        "description": "Email used by user",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/NewUser"
                        },
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "body",
                        "description": "Password used by user",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/NewUser"
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success - respond with a JWT Token"
                    },
                    "409": {
                        "description": "User already exists."
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "NewUser": {
            "required": [
                "name",
                "email",
                "provider"
            ],
            "properties": {
                "name": {
                    "description": "User's full name",
                    "type": "string",
                    "example": "Beltrano da Silva"
                },
                "email": {
                    "description": "User's email",
                    "type": "string",
                    "example": "beltrano@gmail.com"
                },
                "provider": {
                    "description": "User's account provider",
                    "type": "string",
                    "example": [
                        "local",
                        "facebook",
                        "google"
                    ]
                }
            }
        },
...
    }
}